[marathon-lb] Fix defaults, permit image config.
This adds a new config item to change the Docker image for marathon-lb,
and also changes the default role back to `slave_public`. Due to the way
mustache templates work and how Marathon handles roles, it's very
difficult to specify 'no role' as the default without adding an addition
config parameter to wrap the role.